A FPS Playable Character
So today i started an unreal project, i'm going to be making a hack and slash medieval game which i will be using for my dissertation for my final year of university. In this commit i made a character move with the w,a,s,d keys moving forward/backwards | left/right. The mouse allows for rotation up and down too, I have done all this through C++ as its somethign im keen to learn Next time i will be implementing animations to show movement instead of a t-pos mesh