Player Animations
In this commit i made 3 animation transitions with an animation blueprint. This allows the player character stay idle when not moving, Walking around or running. Running is triggered by a key press and walking is triggered by movement in any direction.

Soon i plan to make combat mechanics where other objects can die or be knocked back by a sword and allowing the character to protect themselves with a shield.