AI Can Move
I made sure that the AI Character in the game can chase after the player and also patrol around at a random direction. I also had to edit the collision as there was an issue with the AI Damage logic. When the AI Hit the player capsule collider the AI took damage. So i had to change a few settings in the BPs