Player Health and edits
I edited the players animation Blueprint so there was more smoother logic with the animations. Strafing left and right, also walking backward so the animations did not look odd. I also added some logic to some animations so players cant move during block or death.

While i edited the animations i also edited the damage logic for the NPC and added the players health. To allow for more difficulty i allowed the enemies damage to change and cast this value to the player instead of using a consistent value inside the BP of each individual actor. So instead of reducing the health of the player with a value we set up in the players bp we use the enemies value.

I also made a Game Manager which lets the player score points when the Enemy dies and will store a lot of the UI values to control