Player Pushback and UI
i made the player have pushback when blocking the enmy attack, an annimation can be used for this later on. I also added some UI for the start menu and quit menu.

The next thing will be fixing the players sword collider and the enemies collider. The players sword does not always register and the enemies needs to be disabled and enables so players dont take damagewhen the AI does not attack