NPC and PLayer Health Bars
In this commit i made 2 types of heathbars one for the player which shows on the screen and one for the NPC which resets above the AI Actor. I also fixed the collision to keep it consistent. The overlaps kepts making it that players could do more then the required damage. So i used a DoOnce so the logic would happen until reset. I also made a health pickup and added environment assets. I also fixed the NPC collision where some attacks sent the player flying through the air