AI Attack Distance, Extra Damage Pickup and other edits
In this commit i made it so that the AI could only attack the player within a distance. So that the AI was not so close to the player. I also turned off the enemy weapon collider when itis not arracking as it was clipping the player sometimes and doing random damage. There was also an issue with the player pickups damaging the player so i made them NavMesh obsticles so the AI would avoid them. I also made a pickup that gives the player additional damage to hurt enemies with for 15 seconds. I also edited the players sword collider as it was to small and was not covering the whole blade. i also edited some of the UI and made it so that the pick up would spawn randomly when enemies die